Common renderTable f for the 2 confirmaiton tables ¯\_(ツ)_/¯
