+++ ccm_web/client/src/components/BulkSectionCreateUploadConfirmationTable.tsx
@@ -1,7 +1,5 @@
-import React, { useEffect, useState } from 'react'
-import { Paper, Table, TableBody, TableCell, TableContainer, TableFooter, TableHead, TablePagination, TableRow } from '@material-ui/core'
-import { TablePaginationActions } from './TablePagination'
-import StyledTableCell from './StyledTableCell'
+import { useEffect, useState } from 'react'
+import renderTable from './ConfirmationTableTableRenderer'
 
 interface Section {
   rowNumber: number
@@ -37,73 +35,7 @@ function BulkSectionCreateUploadConfirmationTable (props: BulkSectionCreateUploa
     )
   }, [props.sectionNames])
 
-  const renderTable = (): JSX.Element => {
-    const emptyRows = rowsPerPage - Math.min(rowsPerPage, tableRows.length - page * rowsPerPage)
-
-    const handleChangePage = (event: unknown, newPage: number): void => {
-      setPage(newPage)
-    }
-
-    return (
-      <TableContainer component={Paper}>
-        <Table stickyHeader aria-label="custom pagination table">
-          <TableHead>
-              <TableRow>
-                {columns.map((column) => (
-                  <StyledTableCell
-                    key={column.id}
-                    align={column.align}
-                    style={{ minWidth: column.minWidth }}
-                  >
-                    {column.label}
-                  </StyledTableCell>
-                ))}
-              </TableRow>
-            </TableHead>
-          <TableBody>
-            {tableRows.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage).map((row) => {
-              return (
-                <TableRow hover role="checkbox" tabIndex={-1} key={row.rowNumber}>
-                  {columns.map((column) => {
-                    const value = row[column.id]
-                    return (
-                      <TableCell key={column.id} align={column.align}>
-                        {value}
-                      </TableCell>
-                    )
-                  })}
-                </TableRow>
-              )
-            })}
-            {emptyRows > 0 && (
-              <TableRow style={{ height: 53 * emptyRows }}>
-                <TableCell colSpan={6} />
-              </TableRow>
-            )}
-          </TableBody>
-          <TableFooter>
-            <TableRow>
-              <TablePagination
-                rowsPerPageOptions={[rowsPerPage]}
-                colSpan={3}
-                count={tableRows.length}
-                rowsPerPage={rowsPerPage}
-                page={page}
-                SelectProps={{
-                  inputProps: { 'aria-label': 'rows per page' },
-                  native: true
-                }}
-                onChangePage={handleChangePage}
-                ActionsComponent={TablePaginationActions}
-              />
-            </TableRow>
-          </TableFooter>
-        </Table>
-      </TableContainer>
-    )
-  }
-
-  return renderTable()
+  return renderTable(tableRows, columns, page, setPage, rowsPerPage)
 }
 
 export type { BulkSectionCreateUploadConfirmationTableProps, Section }
